extract ChainListHeader & ChainListFooter components from assets
