+++ src/screens/Assets/collectibles/CollectiblesTab.js
@@ -47,8 +47,8 @@ import type { Chain } from 'models/Chain';
 
 // Local
 import { type FlagPerChain, useExpandItemsPerChain } from '../utils';
-import ChainListHeader from '../components/ChainListHeader';
-import ChainListFooter from '../components/ChainListFooter';
+import ChainListHeader from 'components/modern/ChainListHeader';
+import ChainListFooter from 'components/modern/ChainListFooter';
 import { type CollectibleItem, useCollectibleAssets } from './selectors';
 import { buildCollectibleNavigationParams } from './utils';
 import CollectibleListItem from './CollectibleListItem';
