+++ src/screens/Assets/investments/InvestmentsTab.js
@@ -49,8 +49,8 @@ import type { InvestmentAssetBalance } from 'models/Balances';
 
 // Local
 import { type FlagPerChain, useExpandItemsPerChain } from '../utils';
-import ChainListHeader from '../components/ChainListHeader';
-import ChainListFooter from '../components/ChainListFooter';
+import ChainListHeader from 'components/modern/ChainListHeader';
+import ChainListFooter from 'components/modern/ChainListFooter';
 import ServiceListHeader from '../components/ServiceListHeader';
 import {
   useInvestmentsBalance,
